Revert "Split DrrrClient into multiple files"
This reverts commit 25667c580817a57a5218e358284f4230cb16e42d.

At the moment, I'd prefer all of the Client's "Core functionality" be kept in one place.
